+++ modules/tinymce/src/core/main/ts/keyboard/ContentEndpointNavigation.ts
@@ -2,16 +2,16 @@ import { Arr, Fun } from '@ephox/katamari';
 import { Compare, Insert, PredicateFind, SugarElement, SugarNode } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
+import Schema from '../api/html/Schema';
 import CaretPosition from '../caret/CaretPosition';
 import { isAtFirstLine, isAtLastLine } from '../caret/LineReader';
-import * as ElementType from '../dom/ElementType';
 import * as ForceBlocks from '../ForceBlocks';
 
 const isTarget = (node: SugarElement<Node>) => Arr.contains([ 'figcaption' ], SugarNode.name(node));
 
-const getClosestTargetBlock = (pos: CaretPosition, root: SugarElement<HTMLElement>) => {
+const getClosestTargetBlock = (pos: CaretPosition, root: SugarElement<HTMLElement>, schema: Schema) => {
   const isRoot = Fun.curry(Compare.eq, root);
-  return PredicateFind.closest(SugarElement.fromDom(pos.container()), ElementType.isBlock, isRoot).filter(isTarget);
+  return PredicateFind.closest(SugarElement.fromDom(pos.container()), (el) => schema.isBlock(SugarNode.name(el)), isRoot).filter(isTarget);
 };
 
 const isAtFirstOrLastLine = (root: SugarElement<HTMLElement>, forward: boolean, pos: CaretPosition) =>
@@ -21,7 +21,7 @@ const moveCaretToNewEmptyLine = (editor: Editor, forward: boolean) => {
   const root = SugarElement.fromDom(editor.getBody());
   const pos = CaretPosition.fromRangeStart(editor.selection.getRng());
 
-  return getClosestTargetBlock(pos, root).exists(() => {
+  return getClosestTargetBlock(pos, root, editor.schema).exists(() => {
     if (isAtFirstOrLastLine(root, forward, pos)) {
       const insertFn = forward ? Insert.append : Insert.prepend;
       const rng = ForceBlocks.insertEmptyLine(editor, root, insertFn);