+++ modules/tinymce/src/core/test/ts/browser/dom/DomUtilsTest.ts
@@ -15,6 +15,8 @@ describe('browser.tinymce.core.dom.DOMUtilsTest', () => {
   const getTestElement = (id: string = 'test') =>
     DOM.get(id) as HTMLElement;
 
+  const createSvgElement = (name: string) => document.createElementNS('http://www.w3.org/2000/svg', name);
+
   it('parseStyle', () => {
     DOM.add(document.body, 'div', { id: 'test' });
 
@@ -261,12 +263,35 @@ describe('browser.tinymce.core.dom.DOMUtilsTest', () => {
     assert.equal(DOM.getAttrib(nonElement, 'test', ''), '');
     assert.equal(DOM.getAttrib(nonElement, 'test', 'x'), 'x');
 
+    const svg = createSvgElement('svg');
+    DOM.setAttrib(svg, 'class', 'foo');
+    assert.equal(DOM.getAttrib(svg, 'class'), 'foo');
+
     DOM.remove('test');
   });
 
   it('setGetAttrib on null', () => {
     assert.equal(DOM.getAttrib(null, 'test'), '');
     DOM.setAttrib(null, 'test', null);
+  });
+
+  it('getContentEditable', () => {
+    const test = DOM.add(document.body, 'div', { id: 'test' });
+
+    test.innerHTML = `
+      <span contenteditable="false"></span>
+      <span contenteditable="true"></span>
+      <span></span>
+      <svg></svg>
+    `;
+    const elms = test.children;
+
+    assert.equal(DOM.getContentEditable(elms[0]), 'false');
+    assert.equal(DOM.getContentEditable(elms[1]), 'true');
+    assert.equal(DOM.getContentEditable(elms[2]), null, 'inherit counts as null');
+    assert.equal(DOM.getContentEditable(elms[3]), null, 'We can not get the contentEditable state from SVGElement since it does not have that property');
+
+    DOM.remove('test');
   });
 
   it('getAttribs', () => {
@@ -323,6 +348,12 @@ describe('browser.tinymce.core.dom.DOMUtilsTest', () => {
     assert.isUndefined(DOM.getStyle(null, 'fontSize'));
 
     DOM.remove('test');
+  });
+
+  it('setStyle/getStyle on a SVG element', () => {
+    const svg = createSvgElement('svg');
+    DOM.setStyle(svg, 'color', 'red');
+    assert.equal(DOM.getStyle(svg, 'color'), 'red');
   });
 
   it('getPos', () => {
@@ -933,6 +964,12 @@ describe('browser.tinymce.core.dom.DOMUtilsTest', () => {
       editableRoot: false,
       path: [ 0, 0 ],
       expected: true
+    }));
+
+    it('TINY-10237: isEditable on SVG element should always be considered non-editable', testIsEditable({
+      input: '<svg></svg>',
+      path: [ 0 ],
+      expected: false
     }));
   });
 });