+++ modules/tinymce/src/themes/silver/main/ts/ui/core/color/ColorSwatch.ts
@@ -1,5 +1,5 @@
 import { HexColour, RgbaColour } from '@ephox/acid';
-import { Cell, Fun, Optional, Optionals } from '@ephox/katamari';
+import { Arr, Cell, Fun, Optional, Optionals, Strings } from '@ephox/katamari';
 import { Css, SugarElement, SugarNode, TransformFind } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -122,15 +122,32 @@ const setIconColor = (splitButtonApi: Toolbar.ToolbarSplitButtonInstanceApi | Me
   splitButtonApi.setIconFill(id, newColor);
 };
 
+const setTooltip = (buttonApi: Toolbar.ToolbarSplitButtonInstanceApi | Menu.NestedMenuItemInstanceApi, tooltip: string) => {
+  buttonApi.setTooltip(tooltip);
+};
+
 const select = (editor: Editor, format: ColorFormat) =>
   (value: string) => {
     const optCurrentHex = getCurrentColor(editor, format);
     return Optionals.is(optCurrentHex, value.toUpperCase());
   };
 
-const registerTextColorButton = (editor: Editor, name: string, format: ColorFormat, tooltip: string, lastColor: Cell<string>) => {
+// Selecting `Remove Color` would set the lastColor to ''
+const getToolTipText = (editor: Editor, format: ColorFormat, lastColor: string) => {
+  if (Strings.isEmpty(lastColor)) {
+    return format === 'forecolor' ? 'Text color' : 'Background color';
+  }
+
+  const tooltipPrefix = format === 'forecolor' ? 'Text color {0}' : 'Background color {0}';
+  const colors = getColors(Options.getColors(editor, format), format, false);
+  const colorText = Arr.find(colors, (c) => c.value === lastColor).getOr({ text: '' }).text;
+
+  return editor.translate([ tooltipPrefix, editor.translate(colorText) ]);
+};
+
+const registerTextColorButton = (editor: Editor, name: string, format: ColorFormat, lastColor: Cell<string>) => {
   editor.ui.registry.addSplitButton(name, {
-    tooltip,
+    tooltip: getToolTipText(editor, format, lastColor.get()),
     presets: 'color',
     icon: name === 'forecolor' ? 'text-color' : 'highlight-bg-color',
     select: select(editor, format),
@@ -155,6 +172,7 @@ const registerTextColorButton = (editor: Editor, name: string, format: ColorForm
       const handler = (e: EditorEvent<{ name: string; color: string }>) => {
         if (e.name === name) {
           setIconColor(splitButtonApi, e.name, e.color);
+          setTooltip(splitButtonApi, getToolTipText(editor, format, e.color));
         }
       };
 
@@ -175,6 +193,7 @@ const registerTextColorMenuItem = (editor: Editor, name: string, format: ColorFo
     text,
     icon: name === 'forecolor' ? 'text-color' : 'highlight-bg-color',
     onSetup: (api) => {
+      setTooltip(api, getToolTipText(editor, format, lastColor.get()));
       setIconColor(api, name, lastColor.get());
       return onSetupEditableToggle(editor)(api);
     },
@@ -267,8 +286,8 @@ const register = (editor: Editor): void => {
   const fallbackColorBackground = Options.getDefaultBackgroundColor(editor);
   const lastForeColor = Cell(fallbackColorForeground);
   const lastBackColor = Cell(fallbackColorBackground);
-  registerTextColorButton(editor, 'forecolor', 'forecolor', 'Text color', lastForeColor);
-  registerTextColorButton(editor, 'backcolor', 'hilitecolor', 'Background color', lastBackColor);
+  registerTextColorButton(editor, 'forecolor', 'forecolor', lastForeColor);
+  registerTextColorButton(editor, 'backcolor', 'hilitecolor', lastBackColor);
 
   registerTextColorMenuItem(editor, 'forecolor', 'forecolor', 'Text color', lastForeColor);
   registerTextColorMenuItem(editor, 'backcolor', 'hilitecolor', 'Background color', lastBackColor);