+++ modules/tinymce/src/themes/silver/main/ts/ui/window/SilverDialogCommon.ts
@@ -4,14 +4,13 @@ import {
 } from '@ephox/alloy';
 import { Dialog, DialogManager } from '@ephox/bridge';
 import { Arr, Cell, Obj, Optional } from '@ephox/katamari';
-import { Height, SelectorFind } from '@ephox/sugar';
+import { Class, Classes, Height, SelectorFind, SugarElement } from '@ephox/sugar';
 
 import { UiFactoryBackstage, UiFactoryBackstageProviders } from '../../backstage/Backstage';
 import * as RepresentingConfigs from '../alien/RepresentingConfigs';
 import { StoredMenuButton, StoredMenuItem } from '../button/MenuButton';
 import * as Dialogs from '../dialog/Dialogs';
 import { FormBlockEvent, formCancelEvent } from '../general/FormEvents';
-import { dialogChannel } from './DialogChannels';
 import { ExtraListeners } from './SilverDialogEvents';
 import { renderModalHeader } from './SilverDialogHeader';
 
@@ -70,36 +69,54 @@ const getEventExtras = (lazyDialog: () => AlloyComponent, providers: UiFactoryBa
   }
 });
 
-const renderModalDialog = <T>(spec: DialogSpec, initialData: T, dialogEvents: AlloyEvents.AlloyEventKeyAndHandler<any>[], backstage: UiFactoryBackstage): AlloyComponent => {
-  const updateState = (_comp: AlloyComponent, incoming: T) => Optional.some(incoming);
+const fullscreenClass = 'tox-dialog--fullscreen';
+const largeDialogClass = 'tox-dialog--width-lg';
+const mediumDialogClass = 'tox-dialog--width-md';
 
-  return GuiFactory.build(Dialogs.renderDialog({
-    ...spec,
-    firstTabstop: 1,
-    lazySink: backstage.shared.getSink,
-    extraBehaviours: [
-      // Because this doesn't define `renderComponents`, all this does is update the state.
-      // We use the state for the initialData. The other parts (body etc.) render the
-      // components based on what reflecting receives.
-      Reflecting.config({
-        channel: `${dialogChannel}-${spec.id}`,
-        updateState,
-        initialData
-      }),
-      RepresentingConfigs.memory({ }),
-      ...spec.extraBehaviours
-    ],
-    onEscape: (comp) => {
-      AlloyTriggers.emit(comp, formCancelEvent);
-    },
-    dialogEvents,
-    eventOrder: {
-      [SystemEvents.receive()]: [ Reflecting.name(), Receiving.name() ],
-      [SystemEvents.attachedToDom()]: [ 'scroll-lock', Reflecting.name(), 'messages', 'dialog-events', 'alloy.base.behaviour' ],
-      [SystemEvents.detachedFromDom()]: [ 'alloy.base.behaviour', 'dialog-events', 'messages', Reflecting.name(), 'scroll-lock' ]
-    }
-  }));
+const getDialogSizeClass = (size: Dialog.DialogSize): Optional<string> => {
+  switch (size) {
+    case 'large':
+      return Optional.some(largeDialogClass);
+    case 'medium':
+      return Optional.some(mediumDialogClass);
+    default:
+      return Optional.none();
+  }
 };
+
+const updateDialogSizeClass = (size: Dialog.DialogSize, component: AlloyComponent): void => {
+  const dialogBody = SugarElement.fromDom(component.element.dom);
+  if (!Class.has(dialogBody, fullscreenClass)) {
+    Classes.remove(dialogBody, [ largeDialogClass, mediumDialogClass ]);
+    getDialogSizeClass(size).each((dialogSizeClass) => Class.add(dialogBody, dialogSizeClass));
+  }
+};
+
+const toggleFullscreen = (comp: AlloyComponent, currentSize: Dialog.DialogSize): void => {
+  const dialogBody = SugarElement.fromDom(comp.element.dom);
+  const classes = Classes.get(dialogBody);
+  const currentSizeClass = Arr.find(classes, (c) => c === largeDialogClass || c === mediumDialogClass).or(getDialogSizeClass(currentSize));
+  Classes.toggle(dialogBody, [ fullscreenClass, ...currentSizeClass.toArray() ]);
+};
+
+const renderModalDialog = (spec: DialogSpec, dialogEvents: AlloyEvents.AlloyEventKeyAndHandler<any>[], backstage: UiFactoryBackstage): AlloyComponent => GuiFactory.build(Dialogs.renderDialog({
+  ...spec,
+  firstTabstop: 1,
+  lazySink: backstage.shared.getSink,
+  extraBehaviours: [
+    RepresentingConfigs.memory({ }),
+    ...spec.extraBehaviours
+  ],
+  onEscape: (comp) => {
+    AlloyTriggers.emit(comp, formCancelEvent);
+  },
+  dialogEvents,
+  eventOrder: {
+    [SystemEvents.receive()]: [ Reflecting.name(), Receiving.name() ],
+    [SystemEvents.attachedToDom()]: [ 'scroll-lock', Reflecting.name(), 'messages', 'dialog-events', 'alloy.base.behaviour' ],
+    [SystemEvents.detachedFromDom()]: [ 'alloy.base.behaviour', 'dialog-events', 'messages', Reflecting.name(), 'scroll-lock' ]
+  }
+}));
 
 const mapMenuButtons = (buttons: Dialog.DialogFooterButton[], menuItemStates: Record<string, Cell<boolean>> = {}): (Dialog.DialogFooterButton | StoredMenuButton)[] => {
   const mapItems = (button: Dialog.DialogFooterMenuButton): StoredMenuButton => {
@@ -137,6 +154,9 @@ export {
   getBusySpec,
   getHeader,
   getEventExtras,
+  updateDialogSizeClass,
+  getDialogSizeClass,
+  toggleFullscreen,
   renderModalDialog,
   mapMenuButtons,
   extractCellsToObject